Improve mouse controls
I think good aim assist would be better than mouse controls

+++ src/comps/input.hpp
@@ -14,8 +14,8 @@ struct KeyInput {};
 
 /// If present, the mouse will control MoveInput
 struct MouseInput {
-  float x = 0.0f;
-  float y = 0.0f;
+  int x = 0;
+  int y = 0;
 };
 
 /// If present, the blaster bolts are nudged toward the nearest target
